+++ src/components/Canvas/CanvasOuter.default.tsx
@@ -1,6 +1,8 @@
 import styled from 'styled-components'
+import { IConfig } from '../../types'
 
 export interface ICanvasOuterDefaultProps {
+  config: IConfig
   children: any
   ref?: React.Ref<any>
 }