+++ src/components/NodeInner/NodeInner.default.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react'
 import styled from 'styled-components'
-import { INode } from '../../'
+import { IConfig, INode } from '../../'
 
 export interface INodeInnerDefaultProps {
+  config: IConfig
   node: INode
 }