+++ src/components/Port/Port.wrapper.tsx
@@ -1,7 +1,9 @@
+import { isEqual } from 'lodash'
 import * as React from 'react'
+import * as ReactDOM from 'react-dom'
 import { v4 } from 'uuid'
 import {
-  ILink, INode, IOnLinkCancel, IOnLinkComplete, IOnLinkMove,
+  IConfig, ILink, INode, IOnLinkCancel, IOnLinkComplete, IOnLinkMove,
   IOnLinkStart, IOnPortPositionChange, IPort, IPosition, ISelectedOrHovered,
 } from '../../'
 import CanvasContext from '../Canvas/CanvasContext'
@@ -18,6 +20,7 @@ const composedPath = (el: HTMLElement | null) => {
 }
 
 export interface IPortWrapperProps {
+  config: IConfig
   style?: object
   offset: IPosition
   selected: ISelectedOrHovered | undefined
@@ -40,29 +43,26 @@ export class PortWrapper extends React.Component<IPortWrapperProps> {
   public static contextType = CanvasContext
   public context!: React.ContextType<typeof CanvasContext>
 
-  public nodeRef?: HTMLDivElement
+  private nodeRef = React.createRef<HTMLDivElement>()
 
-  public getNodRef = (el: HTMLDivElement) => {
-    if (el) {
-      const { node, port, onPortPositionChange } = this.props
-      this.nodeRef = el
-
-      // Ports component should be positions absolute
-      // Factor this in so we get position relative to the node
-      const nodesEl = el.parentElement
-        ? el.parentElement
-        : { offsetLeft: 0, offsetTop: 0 }
-
-      const position = {
-        x: el.offsetLeft + nodesEl.offsetLeft + el.offsetWidth / 2,
-        y: el.offsetTop + nodesEl.offsetTop + el.offsetHeight / 2,
+  public componentDidUpdate (prevProps: IPortWrapperProps) {
+    // Update port position after a re-render if node.size has changed
+    if (!isEqual(this.props.node.size, prevProps.node.size)) {
+      const el = ReactDOM.findDOMNode(this.nodeRef.current) as HTMLInputElement
+      if (el) {
+        // Ports component should be positions absolute
+        // Factor this in so we get position relative to the node
+        const nodesEl = el.parentElement
+          ? el.parentElement
+          : { offsetLeft: 0, offsetTop: 0 }
+        // update port position after node size has been determined
+        this.props.onPortPositionChange({ config: this.props.config, node: this.props.node, port: this.props.port, el, nodesEl })
       }
-      onPortPositionChange(node, port, position)
     }
   }
 
   public onMouseDown = (startEvent: React.MouseEvent) => {
-    const { offset, node, port, onLinkStart, onLinkCancel, onLinkComplete, onLinkMove } = this.props
+    const { offset, node, port, onLinkStart, onLinkCancel, onLinkComplete, onLinkMove, config } = this.props
     const linkId = v4()
     const fromNodeId = node.id
     const fromPortId = port.id
@@ -73,6 +73,7 @@ export class PortWrapper extends React.Component<IPortWrapperProps> {
       const { offsetX, offsetY } = this.context
 
       onLinkMove({
+        config,
         linkId, startEvent, fromNodeId, fromPortId,
         toPosition: {
           x: e.clientX - offsetX - offset.x,
@@ -99,9 +100,9 @@ export class PortWrapper extends React.Component<IPortWrapperProps> {
       if (portEl) {
         const toPortId = portEl.getAttribute('data-port-id') as string
         const toNodeId = portEl.getAttribute('data-node-id') as string
-        onLinkComplete({ linkId, startEvent, fromNodeId, fromPortId, toNodeId, toPortId })
+        onLinkComplete({ config, linkId, startEvent, fromNodeId, fromPortId, toNodeId, toPortId })
       } else {
-        onLinkCancel({ linkId, startEvent, fromNodeId, fromPortId })
+        onLinkCancel({ config, linkId, startEvent, fromNodeId, fromPortId })
       }
 
       // Remove the listeners if the link is complete or canceled
@@ -114,7 +115,7 @@ export class PortWrapper extends React.Component<IPortWrapperProps> {
     window.addEventListener('mousemove', mouseMoveHandler, false)
 
     // Notify state of link start
-    onLinkStart({ linkId, startEvent, fromNodeId, fromPortId })
+    onLinkStart({ config, linkId, startEvent, fromNodeId, fromPortId })
 
     // Prevent default and stop propagation to prevent text selection
     startEvent.preventDefault()
@@ -130,6 +131,7 @@ export class PortWrapper extends React.Component<IPortWrapperProps> {
       port,
       node,
       Component = PortDefault,
+      config,
     } = this.props
 
     return (
@@ -137,10 +139,11 @@ export class PortWrapper extends React.Component<IPortWrapperProps> {
         data-port-id={port.id}
         data-node-id={node.id}
         onMouseDown={this.onMouseDown}
-        ref={this.getNodRef}
+        ref={this.nodeRef}
         style={style}
       >
         <Component
+          config={config}
           port={port}
           isSelected={!!selected && selected.type === 'port' && selected.id === port.id}
           isHovered={!!hovered && hovered.type === 'port' && hovered.id === port.id}