+++ src/components/PortsGroup/PortsGroup.default.tsx
@@ -1,6 +1,8 @@
 import styled, { css } from 'styled-components'
+import { IConfig } from '../../'
 
 export interface IPortsGroupDefaultProps {
+  config: IConfig
   side: 'top' | 'bottom' | 'left' | 'right'
 }