+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './chart'
+export * from './config'
 export * from './functions'
 export * from './generics'