+++ stories/index.tsx
@@ -1,8 +1,9 @@
 import { storiesOf } from '@storybook/react'
 import * as React from 'react'
+import { ConfigSnapToGridDemo } from './ConfigSnapToGrid'
+import { ConfigValidateLinkDemo } from './ConfigValidateLink'
 import { CustomCanvasOuterDemo } from './CustomCanvasOuter'
 import { CustomLinkDemo } from './CustomLink'
-import { CustomNodeDemo } from './CustomNode'
 import { CustomNodeInnerDemo } from './CustomNodeInner'
 import { CustomPortDemo } from './CustomPort'
 import { DragAndDropSidebar } from './DragAndDropSidebar'
@@ -15,12 +16,9 @@ import { StressTestDemo } from './StressTest'
 storiesOf('State', module)
   .add('Internal React State', InternalReactState)
   .add('External React State', () => <ExternalReactState />)
-storiesOf('Readonly Mode', module)
-  .add('default', ReadonlyMode)
 
 storiesOf('Custom Components', module)
-  .add('Node', CustomNodeDemo)
-  .add('Node Inner', CustomNodeInnerDemo)
+  .add('Node Inner', () => <CustomNodeInnerDemo />)
   .add('Port', CustomPortDemo)
   .add('Canvas Outer', CustomCanvasOuterDemo)
   .add('Canvas Link', () => <CustomLinkDemo />)
@@ -31,3 +29,8 @@ storiesOf('Stress Testing', module)
 storiesOf('Sidebar', module)
   .add('Drag and Drop', DragAndDropSidebar)
   .add('Selected Sidebar', () => <SelectedSidebar />)
+
+storiesOf('Other Config', module)
+  .add('Snap To Grid', ConfigSnapToGridDemo)
+  .add('Link validation function', ConfigValidateLinkDemo)
+  .add('Read only mode', ReadonlyMode)